+++ matrix-sdk-android/src/main/java/org/matrix/android/sdk/api/session/room/failure/CreateRoomFailure.kt
@@ -18,8 +18,9 @@
 package org.matrix.android.sdk.api.session.room.failure
 
 import org.matrix.android.sdk.api.failure.Failure
+import org.matrix.android.sdk.api.failure.MatrixError
 
 sealed class CreateRoomFailure : Failure.FeatureFailure() {
-
-    object CreatedWithTimeout: CreateRoomFailure()
+    object CreatedWithTimeout : CreateRoomFailure()
+    data class CreatedWithFederationFailure(val matrixError: MatrixError) : CreateRoomFailure()
 }