+++ CHANGES.md
@@ -9,6 +9,7 @@ Improvements 🙌:
  - Give user the possibility to prevent accidental call (#1869)
  - Display device information (name, id and key) in Cryptography setting screen (#1784)
  - Ensure users do not accidentally ignore other users (#1890)
+ - Better handling DM creation when invitees cannot be inviting due to denied federation
  - Support new config.json format and config.domain.json files (#1682)
  - Increase Font size on Calling screen (#1643)
  - Make the user's Avatar live in the general settings